Update: improve README instructions
The README seemed outdated and the "Build Setup"
section did not work if you followed line by
line.

In this commit a "Getting stared" section is added
to make it easier for other developers to setup
the project.

In addition to this some sections are moved/rewritten
to categorize and improve them under "Development" and
"Building".