feat: add Top Stopwatch Events view (#659)
In order for the stopwatch events to appear, one has to enable "Force
devmode" in the settings and have some stopwatch events.

The stopwatch events are merged with window events. If they overlap,
then the stopwatch events have precedence, so that no more than 24 h of
all events are logged per day.