feat: add universal columns vis
If you wanted to visualize a custom watcher, you would need to either
use the timeline view or make your own. This makes a universal vis that
just shows arbitrary keys from the data that is saved. While it might
not be great for some special types of custom watchers, it works really
well for watchers such as the lastfm.